Moved into a model Package
So, the way I configured this, it didn't actually have the model package
in the directory structure. Fixed. xD
