--- UrbanParksProject/src/Volunteer.java
@@ -1,6 +1,0 @@
-package model;
-
-public class Volunteer {
-	public String firstName;
-	public String lastName;
-}
