+++ UrbanParksProject/src/model/Volunteer.java
@@ -1,0 +1,6 @@
+package model;
+
+public class Volunteer {
+	public String firstName;
+	public String lastName;
+}
